requirements.txt
selenium>=4.10.0
webdriver-manager>=4.0.1
pandas>=2.1.0
streamlit>=1.30.0
numpy>=1.25.0
seaborn>=0.12.0
matplotlib>=3.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
openpyxl>=3.1.2